% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_samples.R
\name{bootstrap_samples}
\alias{bootstrap_samples}
\title{Bootstrap Sample from a Population Grouped by Items}
\usage{
bootstrap_samples(
  start = 20,
  stop = 100,
  increase = 5,
  population,
  replace = TRUE,
  nsim = 100,
  grouping_items = NULL
)
}
\arguments{
\item{start}{Sample size for the smallest potential sample}

\item{stop}{Sample size for the largest potential sample}

\item{increase}{Number to increase the sample size with for
each potential sample}

\item{population}{The population data set or the pilot dataset}

\item{replace}{A TRUE/FALSE value to bootstrap with replacement}

\item{nsim}{The number of simulations/samples you want to return}

\item{grouping_items}{The names of columns to group your data
by for the simulation, usually this column is the item column}
}
\value{
\describe{
 \item{samples}{The bootstrapped samples for each tested sample
 size. }
}
}
\description{
This function allows you to bootstrap samples across
various sample sizes when the data (optionally) has
repeated measures items.
}
\examples{

# step 1 create data like what I think I'll get or use your own
pops <- simulate_population(mu = 4, mu_sigma = .2, sigma = 2,
  sigma_sigma = .2, number_items = 30, number_scores = 20,
  smallest_sigma = .02, min_score = 1, max_score = 7, digits = 0)

# step 3 simulate bootstrapped samples
samples <- bootstrap_samples(start = 20, stop = 100,
  increase = 5, population = pops,
  replace = TRUE, grouping_items = NULL)

# notice just 20 items
samples[[1]]

samples <- bootstrap_samples(start = 20, stop = 100,
  increase = 5, population = pops,
  replace = TRUE, grouping_items = "item")

# notice 20 rows per item
samples[[1]]

}
