% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_simple.R
\name{fake_simple}
\alias{fake_simple}
\title{Calculate pseudowords by replacing bigrams/single letters}
\usage{
fake_simple(wordlist)
}
\arguments{
\item{wordlist}{a list of tokens you want to use to create
your pseudowords.}
}
\value{
A dataset of original tokens and suggested pseudowords.
}
\description{
This function takes a list of tokens and returns a list
of potential generated pseudowords by replacing characters
based on legal bigrams within the word list set. Note: you
should check the list, as random generation can generate
a new real token.
}
\examples{
# af_top_sim$fake_cue <- fake_simple(af_top_sim$cue)

}
