% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_wuggy.R
\name{fake_Wuggy}
\alias{fake_Wuggy}
\title{Calculate pseudowords with the Wuggy Algorithm}
\usage{
fake_Wuggy(wordlist, language_hyp, lang, replacewords)
}
\arguments{
\item{wordlist}{A list of valid words from which to
calculate the frequencies of syllables and transition
ngrams from.}

\item{language_hyp}{The language hyphenation you want
to use. You can find them to download at https://hyphenation.org/.
Or check out the ones we used in our package /inst/latex.}

\item{lang}{The two letter language code for the language you
imported for hyphenation.}

\item{replacewords}{A list of tokens you want to use to create
your pseudowords.}
}
\value{
A dataset of original tokens and suggested pseudowords.
}
\description{
This function takes a list of tokens and returns a list
of potential generated pseudowords by using the Wuggy.
Note: you should check the list, as random generation
can generate a new real token. Note: this function is
fairly slow as the word list gets larger.
}
