% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_prime.R
\name{import_prime}
\alias{import_prime}
\title{Import the SPAML Data or Stimuli}
\usage{
import_prime(filename)
}
\arguments{
\item{filename}{Include the filename of the dataset you want to load.
Note: some of the `full_data` datasets are very large and require
a good bit of memory to load.}
}
\value{
a dataset or stimuli list in a dataframe.
}
\description{
This function allows you to import the current datasets available from
the SPAML project. You can use `data("primeData")` to view the
list of available datasets/materials.
}
\examples{
# ar_words <- import_prime("ar_words.csv")

}
\keyword{datasets,}
\keyword{linguistic}
\keyword{metadata,}
\keyword{norms}
