% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primeData-data.R
\docType{data}
\name{primeData}
\alias{primeData}
\title{SPAML Priming Data Download}
\format{
A data frame of links and information about the
SPAML project.

`type`: which type of dataset you would like to view/load.
`filename`: the name of the file of the dataset.
`language`: the language of the stimuli/materials in the data.
`location`: a download link for the most recent data.

Please note: we are currently on v1.0.1 for the data from
the project.
}
\usage{
data(primeData)
}
\description{
Dataset to use the \code{\link{import_prime}} function to
import data or materials from the SPAML project.
}
\keyword{datasets}
