% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{processData}
\alias{processData}
\title{Process labjs SQLite Files}
\usage{
processData(database)
}
\arguments{
\item{database}{The filepath of the database you want to process.}
}
\value{
The experiment data in a long format.
}
\description{
This function is a helper function to transform labjs
SQLite files into a long data file where each trial
is one row.
}
\examples{
# df <- processData("filename")

}
\keyword{datasets,}
\keyword{linguistic}
\keyword{metadata,}
\keyword{norms}
