% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_population.R
\name{simulate_population}
\alias{simulate_population}
\title{Simulate Normal Population Scores for Multiple Items}
\usage{
simulate_population(
  mu = NULL,
  mu_sigma = NULL,
  sigma = NULL,
  sigma_sigma = NULL,
  number_items = NULL,
  number_scores = NULL,
  smallest_sigma = NULL,
  min_score = NULL,
  max_score = NULL,
  digits = NULL
)
}
\arguments{
\item{mu}{The population mean for the items.}

\item{mu_sigma}{The amount of variability for each of the
means (i.e., the standard deviation of the item's means)}

\item{sigma}{The population standard deviation for
each item.}

\item{sigma_sigma}{The standard deviation of the standard
deviations for the items (i.e., heterogeneity)}

\item{number_items}{The number of items to generate}

\item{number_scores}{The number of scores for each item to generate}

\item{smallest_sigma}{The smallest possible standard deviation
for an item that is acceptable (default is sigma_sigma/10)}

\item{min_score}{If you want to truncate scores, what is the
smallest possible score?}

\item{max_score}{If you want to truncate scores, what is the
largest possible score?}

\item{digits}{If you want to round scores, how many digits
should it be rounded to?}
}
\value{
\describe{
 \item{"population"}{The population scores with the number of items
and scores specified}
}
}
\description{
This function allows you to create normal populations
for data that would include repeated measures items.
Additionally, the data can be rounded and/or truncated
to ensure it matches a target scale - for example, a
1-7 type rating scale.
}
\examples{
simulate_population(mu = 4, mu_sigma = .2, sigma = 2,
 sigma_sigma = .2, number_items = 30, number_scores = 1000,
 smallest_sigma = .02, min_score = 1, max_score = 7,
 digits = 0)
}
