% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n_similarity.R
\name{top_n}
\alias{top_n}
\title{Calculate the Top N Cosines}
\usage{
top_n(similarity, n)
}
\arguments{
\item{similarity}{the matrix of cosine values.}

\item{n}{the number of most related cosine values you want
to return. Remember that you will get token-token cosine back
as a value of 1, so use n+1 to get the number you want.}
}
\value{
A dataframe of token-token-cosine values.
}
\description{
This function takes a matrix of cosine similarity values,
where the row and column names are tokens. The values will
be calculated for the columns (so if you have a non-square)
matrix, make sure the columns are the values you want.
}
\examples{
# af_top_sim <- top_n(af_cosine, 6)

}
