% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_lab.R
\name{import_lab}
\alias{import_lab}
\title{Import Datasets from the Linguistic Annotated Bibliography}
\usage{
import_lab(bibtexID = NULL, citation = NULL, language = NULL, variables = NULL)
}
\arguments{
\item{bibtexID}{The bibtex ID of the dataset you are trying to load.
You can leave all parameters blank to load just the metadata.}

\item{citation}{Include the citation for the dataset you loaded - will only
load if you include a bibtex ID, use `TRUE` to get the citation.}

\item{language}{If you include a bibtex ID, you will get back the language of
the dataset, if you do not include a bibtex ID, it will return a list of
datasets in that language.}

\item{variables}{If you include a bibtex ID, you will get back the variables
included the dataset, if you do not include a bibtex ID, it will return a list of
datasets that include that variable (can also be paired with language).
Use the column names from the metadata as your filter.}
}
\value{
\itemize{
 \item{metadata}{The metadata list of avaliable datasets}
 \item{loaded_data}{The dataset you requested to load}
 \item{language}{The language of the dataset you requested to load}
 \item{variables}{The variables of the dataset you requested to load}
 \item{datasets}{Possible datasets based on your language
 and variable names}
}
}
\description{
This function allows you to import the current datasets available from
\href{https://github.com/orgs/SemanticPriming/}{Semantic Priming GitHub
Group}. These datasets generally have psycholinguistic variables like
Age of Acquisition, length, neighborhoods, Familiarity, Concretenes,
and more. We will update datasets as we have them. All data belongs
to the person who published them.
}
\examples{

# import_lab()
# import_lab(bibtexID = "Birchenough2017", citation = TRUE)
# import_lab(language = "English", variables = c("aoa", "freq"))
}
\keyword{datasets,}
\keyword{linguistic}
\keyword{metadata,}
\keyword{norms}
